vmwgfx: use common pipe_screen ref counting
Use the common pipe_screen ref counting and fd hashing functions.
The mutex can be dropped as the pipe loader serializes the
create_screen() and destroy() calls.

This changes the reference counting from the winsys to the pipe screen.
Previously, it was possible to have 1 winsys with multiple pipe screens.
Now there will only be a single winsys and pipe screen.

Signed-off-by: Rob Herring <robh@kernel.org>
Cc: Brian Paul <brianp@vmware.com>