gallium: add common pipe_screen reference counting functions
In order to prevent multiple pipe_screens being created in the same
process, lookup of the DRM FD and reference counting of the pipe_screen
are needed. Several implementations of this exist in various gallium
drivers/winsys already. This creates a common version which is opt-in
for winsys implementations.

The callers of pipe_screen_{un}reference() functions are responsible for
serializing the calls. Currently, this is done by the pipe-loader mutex.

Signed-off-by: Rob Herring <robh@kernel.org>