pipe-loader: serialize create_screen() calls with a mutex
Creating a screen needs to be serialized in order to support reusing
existing screen. With this, driver private mutexes in create_screen()
functions can be removed.

pipe_loader_create_screen is made an exported symbol to ensure the same
library function (and mutex) is used when multiple libraries are loaded.

Signed-off-by: Rob Herring <robh@kernel.org>
Cc: Emil Velikov <emil.l.velikov@gmail.com>